Moved the LLMChain and Vector Store interactions under a new /infra/llm folder, updated the clearAI script
